chore: conbine files to slice.go
